feat: 가계부 내역에서 카테고리별 금액 합계 계산 로직 구현 - 총합(수입+지출), 수입, 지출 합계 계산 - 카테고리 코드 null 데이터 제외

+++ src/main/java/com/moneymanager/repository/ledger/LedgerRepository.java
@@ -217,10 +217,10 @@ public class LedgerRepository {
 	}
 
 
-	public Integer count() {
+	public Long count() {
 		String sql = "SELECT COUNT(*) FROM ledger";
 
-		return jdbcTemplate.queryForObject(sql, Integer.class);
+		return jdbcTemplate.queryForObject(sql, Long.class);
 	}
